travis: temporarily blacklist a specific commit in the  `scripted-diff` filter
Commit f471a3be00c2b6433b8c258b716982c0539da13f currently has the text "scripted diff"
rather than "scripted-diff", which causes the scripted-diff linter to fail, saying
that it sees a script but not the magic words.

Changing the linter to detect "scripted diff" causes it to fail on this commit,
because the included script does not work. (For reference, the script is

    r() { sed -i 's/vout must be positive/vout cannot be negative/g' $1 }
    r $(git grep -l 'vout must be positive')

Bet you can't guess why bash rejects this without running it!)

We should remember to change this back.