Use git --no-pager in ci script.
When running 'git log' in a ci script, pass --no-pager to ensure
that a pager (i.e. 'less') is never used. When running in CI,
there is no controlling terminal, so a pager will never be used;
but when running the CI scripts by hand to reproduce CI issues,
the automatic invocation of a pager by 'git log' breaks the script.