Bug fix: when CheckChallenge(...) is called on a trimmed header, temporarily untrim it in order to extract the proof.
Without this fix, there would be an exception in this scenario, because a trimmed
header no longer has any proof information attached to it.

This scenario was occuring when a pre-dynafed node with trimmed blocks would receive
a new block header from its peer. The pre-dynafed node would try to validate the header
by comparing its proof to the previous block's proof, and this would cause an exception
if the previous block had pruned.

New headers are not pruned as soon as they're received. Rather, they are pruned whenever the block index
is flushed to disk. So, this behavior was only being observed in functional tests, when a pre-dynafed trimmed
node was restarted.