coinselection: fix assert error in getselectionwaste
fix for intermittent assertion error in getselectionwaste
introduced in f46275d011c057b915c1129bde4973d5c968d088 as part of merge
for bitcoin/bitcoin#25647

the assertion would trigger in a fundrawtransaction call in
feature_confidential_transactions.py where the transaction inputs
included a non-policy asset.

during the loop in getselectionwaste, if the non-policy input waste was
calculated first the assertion would be triggered because
selected_effective_value has not yet met the selection target, because
non-policy coins have 0 effective value

the fix changed the assertion to an if statement, only adding to the
waste metric when selected_effective_value meets the target.