syscall: whitelist network access for msghand thread
Unlike in Bitcoin, our message handling thread needs to open sockets in
order to call out to the mainchain RPC interface. We recently brought in
a seccomp syscall whitelist from upstream, which is enabled on my local
CI box (though apparently not on Github CI) and it is failing on this.